backend/requirements.txt
# Core Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database (SQLite built-in Python, SQLAlchemy untuk ORM)
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Face Recognition (Sederhana & Powerful)
face-recognition==1.3.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
# dlib==19.24.2  # Uncomment if needed (requires CMake)

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2

# Export & Reports
openpyxl==3.1.2
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
